pyproject.toml
[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[project]
name = "datacommons-mcp-server"
version = "1.0.0"
description = "DataCommons MCP Server for statistical data access"
authors = [
    {name = "Abdullah", email = "abdullah@example.com"}
]

requires-python = ">=3.11,<3.13"
dependencies = [
    "datacommons-mcp>=1.0.0",
    "smithery>=0.4.2",
]

[project.scripts]
dev = "smithery.cli.dev:main" # Run the MCP server in development mode
start = "smithery.cli.start:main" # Run the MCP server in production mode
playground = "smithery.cli.playground:main" # Run server with interactive testing playground

[tool.smithery]
server = "datacommons_mcp_server.server:create_server" # Function that returns a patched FastMCP server
log_level = "warning" # Production log level (optional, defaults to "warning")